move wing models to separate file
